refactor: move home and event registration data fetching to server components
